Clearing view tags for error action
